frontend:fix code review problem
