+++ src/frontend/src/app/shared/packages/kubernetes/create-edit-resource-template.ts
@@ -34,8 +34,8 @@ export class CreateEditResourceTemplate {
 
 
   constructor(
-    private templateService: any,
-    private resourceService: any,
+    public templateService: any,
+    public resourceService: any,
     public location: Location,
     public router: Router,
     public appService: AppService,
@@ -63,33 +63,7 @@ export class CreateEditResourceTemplate {
 
   // 监听提交表单的事件
   onSubmit() {
-    if (this.isSubmitOnGoing) {
-      return;
-    }
-    this.isSubmitOnGoing = true;
-
-    let resourceObj = JSON.parse(JSON.stringify(this.kubeResource));
-    resourceObj = this.generateResource(resourceObj);
-    this.template.ingressId = this.resource.id;
-    this.template.template = JSON.stringify(resourceObj);
-
-    this.template.id = undefined;
-    this.template.name = this.resource.name;
-    this.templateService.create(this.template, this.app.id).subscribe(
-      status => {
-        this.isSubmitOnGoing = false;
-        this.router.navigate(
-          [`portal/namespace/${this.cacheService.namespaceId}/app/${this.app.id}/${this.resourceType}/${this.resource.id}`]
-        );
-        // TODO 路由变化后下面不会生效
-        this.messageHandlerService.showSuccess('创建' + this.resourceType + '模板成功！');
-      },
-      error => {
-        this.isSubmitOnGoing = false;
-        this.messageHandlerService.handleError(error);
-
-      }
-    );
+    // TODO 后续重构可以通过修改类型字段的形式实现该函数的抽象，暂时无法提炼通用
   }
 
   // 监听打开 Modal
